Upgrade to a more recent version of BOSH
Resolves the following CVE:

- https://www.cloudfoundry.org/blog/cve-2018-11083/

This is the smallest version bump to mitigate the CVE.

Previously we had been using compiled BOSH releases, presumably because
it's quicker to cut out the compilation step if possible. However, a
non-vulnerable, compiled v266.x version of BOSH does not appear to be
available. We will have to use the non-compiled version instead.