Ensure the ACM DNS verification record is right
Because some reason it happens that in some environments (specially
prod) the DNS record to verify the certificate is wrong, pointing
to a loadbalancer instead to the right verification value.

To fix this automatically, we change the create_acm_cert.sh script
to also verify if the value of the DNS is right, and update it
if not.