Migrate vars store and clear runtime config
We want to move bosh-dns to the in-manifest addons. This means
converging the Bosh vars store with the original CF vars store. This
effectively reverts the work done in [1].

We use `bosh interpolate` for the YAML tampering because it has built-in
mechanisms for precisely what we want. We essentially just pretend the
CF vars store is a manifest, remove the old values with an ops file,
then source them from the Bosh vars store.

The clearance of the runtime config was done in the same manner as we
have done before (in our Git history).

[1] https://github.com/alphagov/paas-bootstrap/pull/182