Remove Bosh DNS override in Concourse manifest
We had to put this in because we added Bosh DNS to the runtime config,
but didn't want to use it for the Concourse deployment. Now we are
moving Bosh DNS back into the in-manifest addons, so it won't be part of
the Concourse deployment.

We are moving bosh-dns into the manifest in order to enable certificate
rotation and monitoring using our existing patterns. As a result we no
longer need the runtime-config to be present in the codebase.