Add a user for prometheus bosh-exporter
We want to integrate prometheus and use the bosh exporter[1]
to get metrics from bosh and enable auto discovery.

Bosh exporter requires a user to connect. We do not want to use
the admin user directly, so we add a "prometheus" user.

This user will still be an admin user, as all the local users
have admin permissions. But having a separated user would
make it easier convert it later to a read-only user, either
by using UAA or changing the bosh behaviour[2]

[1] https://github.com/bosh-prometheus/bosh_exporter
[2] https://github.com/cloudfoundry/bosh/pull/2068