Fix path to gpg-keys files.
This was erroring when the pieplines were uploaded from a directory
other than the root directory. This manifested when creating a bootstrap
instance as this is invoked from the vagrant directory.

This updates the paths to be relative to the script dir so that it's no
longer dependent on where this is invoked from.