Set BOSH_CLIENT* vars when using bosh
We want to start using bosh with UAA. But using it seems to have
problem when we do a login first with `bosh login` and then
do not pass the `client` and `client-secret` to every command.

To workaround this, we decided to set the variables `BOSH_CLIENT` and
`BOSH_CLIENT_SECRET` in every job that uses bosh.

We replace the script `bosh_login.sh` with the corresponding
snippet that set the environment variables.

Signed-off-by: henrytk <henry.knott@digital.cabinet-office.gov.uk>
Signed-off-by: Hector Rivas Gandara <keymon@gmail.com>