Add db-admin/init-db to create the UAA db in RDS
We want to share the same RDS instance for both BOSH and UAA, but
using different databases.

The AWS primitives to create a RDS instance only allow the
creation of one Database in the RDS service.

The challenge is that during the bootstrap of the environment there is
no VM running in the same VPC, and the bootstrap concourse runs
from a VM outside of the VPC and, hence, without direct access
to postgres.

In order to be able to initially create the UAA database, we
decided create a new bosh-release with a job that would create
databased and roles based on some configuration: db-admin/init-db[1]

We deploy this job colocated with the rest of services in the
bosh VM, and it will race with UAA to create the database.
UAA might fail the first start, but eventually monit will restart
it after the database has been created.

[1] https://github.com/alphagov/paas-db-admin-boshrelease/blob/master/jobs/init-db/spec

Signed-off-by: Hector Rivas Gandara <keymon@gmail.com>
Signed-off-by: henrytk <henry.knott@digital.cabinet-office.gov.uk>