Use a SSH tunnel to connect to bosh from concourse
Bosh API now exposes the url of UAA, which is an internal address.

In order to allow bootstrap concourse to connect to the BOSH and UAA
from an external IP, we will use a SSH tunnel that creates a
SOCKS proxy. We add a script to be source that starts the SSH
command. A trap to stop the tunnel is required to ensure that the
process ends after the task is finished, otherwise concourse hangs
forever.

The bootstrap concourse would add its own external IP as a whitelisted
IP to ssh (vpc terraform, office IPs), so it can reach the SSH port.

This is only required in the bootstrap pipelines that run on the
bootstrap concourse. The deploy-cloudfoundry would always run from
the internal concourse that have direct access to bosh via the VPC.

As we use a SSH tunnel, the paas bootstrap does not need to connect
directly to port 25555, so we remove the rule in the securiry group.

Signed-off-by: Hector Rivas Gandara <keymon@gmail.com>