Deploy UAA in the BOSH VM instance
We want to add UAA as the authentication provider for bosh, so
that we can have more fine grained control of permissions
(e.g read-only users) and we can use federated services like
credhub.

We add the UAA release and configuration inspired in the opsfile
in the bosh-deployment[1].

We only setup the following clients:
 - hm: For the local Health Monitor service.
 - admin: replaces the local admin user

We do not consider the user `admin` and clients `bosh_cli` and
`uaa_admin` defined in the bosh-deployment are required, so we
omitted them for the time being.

Signed-off-by: henrytk <henry.knott@digital.cabinet-office.gov.uk>
Signed-off-by: Hector Rivas Gandara <keymon@gmail.com>