Deploy patched version of bosh v268.2.0
Bosh v268.2.0 seems to have a bug that occassionally causes bosh
dns to wipe DNS entries during a deployment.

This caused API availability issues in our staging deployment.

Reported upstream in slack[1], they claim that there is a bug related
to a patch[2].

Here we use a patched version of bosh including that patch.

[1] https://cloudfoundry.slack.com/archives/C02HPPYQ2/p1542383747442400
[2] https://github.com/cloudfoundry/bosh/commit/bc4bc20fcf8fec20c975d8d6b1bf0c2