Use Terraform to create ACM certs.
Now that the AWS terraform provider supports creating ACM certs, we no
longer need to script this separately, and can mange it using Terraform
which is much more robust.

As part of this change, we no longer generate the apps cert in this
repo. It's only needed by the CF deployment, so doesn't make sense to
include here - we'll manage it with Terraform in the paas-cf repo. It
was originally added here to avoid having to duplicate the scripts into
paas-cf.