Use custom release for Concourse
In the latest Concourse 4.2.1 release the Prometheus metrics exporter doesn't add the job label to the exposed metrics, so we can't alert on individual job behaviours. We've created [1], but we don't know when it's going to be released. We created our own patched version, so we'll use that.

For the changes see [2].

[1] https://github.com/concourse/concourse/pull/2785
[2] https://github.com/alphagov/paas-concourse-bosh-release/compare/v4.2.1..gds_master