bosh-deployment: use custom patched version of bosh
See 3a7ba1443dc24ea73960df100bfd9788aa6cfff6

We update the test to explicitelly indicate the downgrade.

The version 268.4.0 should be fine and includes the patch for the
corresponding bug. The only reason to pin this version is to not
upgrade any software at the same time we are migrating to
bosh-deployment manifests. This should be reverted in a follow up
story.