bosh-deployment: migrate secrets to vars-store
We are going to start using a vars-store.yml file and rely on the
secret and certificate generation from `bosh interpolate`.

But we want to keep the same old secrets and certs in the existing
deployments. We add a script that would be able to populate
a vars-store.yml file from `bosh_secrets.yml` and a directory
with certificates.

We temporary add the secrets to generate-bosh-secrets.rb so
they do not get removed when we run bosh-generate-secrets.