bosh-deployment: secret rotation with vars-store
Now we have migrated the password/certificate generation to
the `bosh interpolate`, we need to rotate the passwords, certificates
and keys in cf-vars-store.

We will use the same script than in paas-cf to rotate these secrets.

We only rotate here the leaf certificates, as we currently don't have
any automated process to rotate the bosh CA.

The old bosh-secrets can be just wiped, as it only contains the
bosh RDS DB password.

Here we copy and duplicate the script rotate-vars-store-secrets.rb
from paas-cf[1], as we do not want to pull the entire paas-cf project
only for this script, and we do not want to extract the script to
it's own project. Eventually, this script will be removed from
paas-cf when we implement secret management using credhub.

[1] https://github.com/alphagov/paas-cf/blob/b0bd5cbc5bfca4ef7f08b45d03d546edaa03d2cd/manifests/cf-manifest/scripts/rotate-vars-store-secrets.rb