bosh-deployment: update versions for bosh-aws-cpi
Set the bosh-release versions for these components as the ones
we have currently.

For bosh-aws-cpi we downgrade to 72 from 73. In this version change
there is a big refactoring[1], and we do not want to upgrade at the
same time we migrate to bosh-deployment. We addthe corresponding
exception in the manifest spec.

[1] https://github.com/cloudfoundry/bosh-aws-cpi-release/compare/v72...v73