Move concourse manifest generation logic out of pipeline
Having the logic in the pipeline meant that it had to be duplicated in
the test helpers, and manually kept in sync. Moving the logic into the
generation script means that the tests can call the same script and have
a much greater change of testing equivelent behaviour.

It also makes it easier to reason about how the concourse manifest is
constructed.