Disable the Bosh compiled package cache.
This feature was used to help in dev environments by allowing
environments to share the compilation outputs thus avoiding some time
compiling things when we update boshrelease versions etc.

While this sounds advantageous, we've hit issues with it relating to how
it identifies the stemcell used in the cache keys, which results in
packages compiled for trusty being used on xenial stemcells, which
causes errors in some situations - see [1]. The discussion on that issue
revealed that this is a feature that upstream are not very interested in
maintaining. We've therefore decided that we should turn it off as the
benefits it brings aren't huge when assessed against the problems we're
having (eg we hit [2] while doing the last 2 CF upgrades, and had to
fork CAPI release).

This removes the only usage of region-specific variables in the bosh
manifest, so removes them as well.

[1]https://github.com/cloudfoundry/bosh/issues/2118
[2]https://github.com/cloudfoundry/cf-deployment/issues/663