Enable UAA http port for healthchecks
In the upgrade from UAA 66 to 68, Tomcat is upgraded from 8.x to 9.x.

This changed the way that Tomcat handled the erroneous situation where a
client sends an HTTP request to an HTTPS connector:

https://github.com/apache/tomcat/commit/3dac1578e9a17ce81239d3b95dab40938489deb2

Previously Tomcat would respond with binary garbage (the first few bytes
of the TLS handshake, probably). Now it responds with an HTTP 400.

It turns out that monit's healthcheck was actually happy with the binary
garbage it was getting before, but it's not happy with the HTTP 400 it
gets back now. That caused the healthcheck to fail consistently, and
eventually monit would kill UAA. Since it had no way of becoming
"healthy" it ended up in a restart loop.

In this commit we tell UAA to listen for HTTP on 8080. The uaa-release
knows to use this port for the healtcheck instead of the HTTPS port.
The HTTP port responds to Monit's HTTP healtchecks properly, so UAA
doesn't get killed anymore.

See https://github.com/cloudfoundry/uaa-release/issues/101