Increase timeouts on all persistent concourse instances
This commit increases the timeout on all persistent concourse environments to 18 hours to match CI. This should make support easier when investigating as there is not the need to constantly log in and lose where you were in the flow. We introduced this after an incident in alpha that resulted in the prod environment being accidentally delete. However we have mitigated this by not pushing the delete pipeline into persistent environments.