Stemcell upgrade to 170.38 as part of cf7.5.0 upgrade
Whilst cf7.5.0 includes stemcell 250.x, we're choosing not to do a major
stemcell bump as part of this upgrade. This is partly due to the non-trivial
BOSH-related issues we've seen `create-env`'ing on top of 250.x, but also
because both 170 and 250 are (a) Xenial-based and (b) supported and fully
patched against known CVEs.

The 250 upgrade will be created as a follow-up story after the cf7.5.0 upgrade.