Retry the concourse-deploy::deploy-concourse step once
All Concourse-VM-modifying pipeline runs will produce an error, as the worker
redeploys itself and the task step errors.

Once a newly-minted worker comes back up, the manual operator action is
*always* to re-run the concourse-deploy job. Adding `attempts: 2` to this
specific step which errors will:

a) result in a green pipeline without operator intervention and
b) run more quickly, as only the step, not the entire job, is rerun.