add multiple concourse workers
context:

we are running into some bad behaviour with scheduling and building
particularly on our build concourse. before this commit we ran worker
and web on the same instances and had only once instance. this commit
enables us to have an arbitrary number of workers, in addition to the
one which runs on the web instance.

this is perhaps a partial step towards multiaz concourse (but this
commit is not a full solution)

---

There are a couple of weird things in this commit

- use --var when doing bosh deploy
- security group "-nocycle"

--var is because spruce renders the value as a string instead of a
number, which is then rejected by bosh because it wants a number. we
don't need spruce to do anything. Spruce ignores ((var)) without spaces
around the parens

-nocycle security groups is because we use ingress/egress rules inline.
we cant add the rules we need inline because it creates a cycle which
terraform is defeated by

I updated the spruce version to get the ignory behaviour, we were
building this version anyway

---

I do not think that we need more instances for deployer, so the number
of worker instances to give it is 0, however for build i chose to give
it two instances which should speed up builds and/or make them more
reliable

Signed-off-by: Toby Lorne <toby.lornewelch-richards@digital.cabinet-office.gov.uk>