Default deployer concourse workers to 1
Build concourses have 3 workers by default, 2 running on their own
instances and 1 colocated with the web node.

Deployer concourses used to have 1 worker by default, only colocated
with the web node.

We get concourse to talk to itself, but currently the way it does this
is using the static IP of the concourse web instance itself as part of
the security group safelist. In order to ensure that the
self-update-pipeline job ran in the right place we added tags which we
can use to select individual instances.

The unfortunate consequence of this is that for deployer concourses
which do not have an external node, resources will not be checked, as
concourse will not allocate untagged resources to tagged nodes.

This commit defaults the deployer concourse to have 1 external worker by
default, which is where all of the actual work will happen, and only use
the colocated-with-web node for updating itself.

Signed-off-by: Toby Lorne <toby.lornewelch-richards@digital.cabinet-office.gov.uk>