Remove ops files removing mbus TLS; override SAN
Since
-> https://www.pivotaltracker.com/n/projects/2197707/stories/162291567
and the corresponding commit
-> 899433d92f8628d38d1c7d5d626c29184e35d9e6
in github.com/cloudfoundry/bosh-agent, MBUS using TLS is mandatory.
Before it would silently use the key/cert from the stemcell, but now it
needs bosh to generate it beforehand

The default external generated cert has the IP as the subject alt name
(SAN), whereas we want the external fqdn

We need to ensure that mbus_bootstrap_ssl contains the correct SAN,
which bosh will not do automatically, so we should clean it before we
generate the manifest. We do this using a new script

Signed-off-by: Toby Lorne <toby.lornewelch-richards@digital.cabinet-office.gov.uk>