Generate and migrate cloud config safely
Context

In order to migrate away from our Concourse BOSH deployment having v1
manifest properties, we need to have a cloud config when bootstrapping.

We need to migrate away frmo manifest properties v1 in order to upgrade
the version of BOSH which we are using.

Prior to this commit this is a problem, because the default cloud config
is set during a run of create-cloudfoundry which is done in paas-cf.

We could move all cloud config into paas-bootstrap but that is also
problematic because the cloud config generated in paas-cf knows about
cloudfoundry terraform, which is only run during create-cloudfoundry.

Fortunately BOSH >=v269 allows you to specify multiple "generic" configs
of any type (runtime, cloud, etc), which are then safely merged
together. Although you, as a developer/operator have to make sure that
you do not have conflicting properties/names.

This means we can have two sets of cloud config:
  - paas-bootstrap
  - paas-cf
which are defined in their respective repos, and set in the respective
pipelines:
  - create-bosh-concourse
  - create-cloudfoundry

This commit does three things:
  - Adds a template cloud config for paas-bootstrap
  - Adds pipeline tasks for generating/updating the paas-bootstrap cloud
  config
  - Adds a temporary task for migrating the default cloud config
  currently generated by paas-cf

---

Migration

The migration is safe but nuanced, and there is an equivalent operation
that has to be performed in paas-cf, unless we want to introduce some
manual operator actual during the migration, which we do not want.

BOSH lets us have multiple cloud configs, but some properties conflict,
and cause safe but annoying runtime errors. PaaS CF currently defines
the vm_type with name "compilation" and the compilation section.

When bootstrapping a new environment, the compilation section needs to
be present in the cloud config, which means it needs to be in
paas-bootstrap's cloud config.

As of this commit:

When the create-bosh-concourse pipeline is run and it sees that there is
a default cloud config, it then:

- Downloads the default cloud config
- Runs two operations against it
  - Remove the "compilation" section
  - Remove the vm_type with name="compilation"
- Re-uploads it as generic config name=paas-cf type=cloud

When the create-bosh-concourse pipeline is run and it does not see
default cloud config, it does not do anything to paas-cf's cloud config

Once paas-cf has been updated to do the right thing, then we can remove
this migration step.

---

Cloud config

The actual cloud config added is a port of the v1 manifest properties in
the concourse bosh manifest; with a few of exceptions:

Compilation

The compilation block and vm_type has been almost verbatim copied from
paas-cf's cloud config, however the network/az is changed from "cf"
which is a network defined only in paas-cf to a new network created here
specifically for compilation, which uses az3.

Because the az names are defined in cloud config, they must not conflict
between paas-cf and paas-bootstrap. I've chosen to define the three
previously implicit AZs in paas-bootstrap as b1, b2, b3 (b for
bootstrap). Despite the terraform outputs having the same names as the
ones in paas-cf, the azs defined here represent different subnets, so we
actually care about 6 BOSH azs (although still only 3 azs in AWS)

Compilation VM needs a public IP when we put it in a subnet without a
NAT gateway which we are doing now

Concourse vm_types and vm_extensions

In BOSH manifest v1 there was a resource_pool specification which has
been decomposed in cloud config to be vm_extensions and vm_types. The
individual resource_pools have been decomposed accordingly

Concourse networks

The concourse bosh manifest has manually defined networks (which we are
not altering as part of this commit/release), however going forward our
cloud config needs to define some azs and networks.

The AZs have been explained above, but the choice of networks here has
not yet been explained.

Previous versions of BOSH and CPIs did not support dynamic networks,
where a dynamic network means IaaS will do IP allocation. As a result
our concourse network needed to be defined manually. The unfortunate
consequence of this is that BOSH would choose an IP already chosen by
Amazon. Making the network dynamic means BOSH will delegate dynamic IP
selection to the IaaS, which is the desired behaviour. BOSH and the
Concourse ELB should no longer fight over 10.0.0.21 which was the
behaviour before.

The consequence of this is that we will not be able to specify the
internal static IP address of Concourse when we switch the Concouse BOSH
manifest to v2 from v1, so that it uses the networks defined in the
paas-boostrap cloud config. I cannot find any dependence on 10.0.0.10 in
either paas-bootstrap or paas-cf's codebase. We rely on a static
external IP in some places, but never the internal one.

---

Please note that this commit does not change the concourse manifest at
all, Concourse is still using the v1 manifest, and its deployment will
not change at all, other than BOSH being aware of the cloud config -
which will then be ignored because of the V1 manifest properties.

Signed-off-by: Toby Lorne <toby.lornewelch-richards@digital.cabinet-office.gov.uk>