Upload concourse secrets to credhub
Generate the Concourse secrets using BOSH (unless we are bootstrapping
in which case hardcode the concourse_web_password.

Use Concourse secrets from BOSH

 We should never be interacting Concourse using the admin
username/password from our laptops unless we are bootstrapping, so just
log in using the browser (OAuth flow).

There are some interesting interactions here with the bootstrapping
process.

We don't self-update-pipelines in bootstrap mode, but in order
to get the pipeline to look the same we want to use ((syntax)) for
concourse web password. We can try and fetch the variable, and if it
fails, continue if we are bootstrapping.

Signed-off-by: Toby Lorne <toby.lornewelch-richards@digital.cabinet-office.gov.uk>