Delete default cloud config
Currently paas-cf sets cloud config which has a the compilation vm
defined, and the compilation section defined, which must be in
the paas-bootstrap cloud config.

In the pipeline, if the default cloud config exists, delete it

A separate PR in paas-cf changes the cloud config to do the right thing:
i.e. not set compilation, and to have a descriptive name

Signed-off-by: Toby Lorne <toby.lornewelch-richards@digital.cabinet-office.gov.uk>