Correct Toby's username in SSH key config
Toby has a long name. He uses "tlwr" in most places, but the convention for
machine usernames is the persons full name as one word. The usernames in the
SSH key config have to match that of the user's machine, so we have to use
Toby's full name here too.