Create UAA (bosh) users for paas-trusted-people
The file in S3 will be uploaded by the `make ENV upload-paas-trusted-people`
make target.

We're generating an ops file which will ultimately add these users to:

```
/var/vcap/jobs/uaa/config/uaa.yml
```

Which is UAA's config file on the bosh director.

Using an ops file is nice because we don't have to think about removing
users who shouldn't have access (if we did this using UAA's API then it
would be more complicated).