Update bosh-cli to use SSO to authenticate to bosh
This turned out to be a bit of a nightmare, because UAA isn't exposed to
the internet - we access it through a SOCKS proxy through an SSH tunnel.

Because the SSO flow requires you to access UAA in a browser, follow the
redirect to google and back again, this means we have to tell the
browser to use our SOCKS proxy.

Most developers use Google Chrome on macOs. This reads its proxy
settings from the system settings. We've updated the bosh-cli script to
temporarily change the system settings (with the user's consent, and
cleaning them up when it exits) to point at a PAC (proxy auto config)
server. This looks like it's good enough to get Chrome and Firefox to
use the SOCKS proxy.

We were also able to get rid of the dependency on docker - now that
bosh-cli is a go binary that's very easy to install there's really no
need for docker.