Use default SSH key when SSHing to Concourse
Some of the team use SSH keys on disk to access Concourse, but others
use keys on a Yubikey. The existing script didn't support using a
Yubikey.

This commit removes all specifying of the SSH key, relying on the
default. It also removes `-o IdentitiesOnly=yes` which was obstructing
Yubikey use.