Shrink deployer-concourse to m5.large in dev
Each team member runs an always-on deployer concourse consisting of a
web node and a worker node, each of which is an m5.xlarge instance.
If each m5.xlarge costs $0.222/hour (current london price; ireland is
slightly cheaper), this corresponds to an on-demand cost of
$0.222*2*24*30 = $319.68 per 30-day month, per developer concourse.

By shrinking the instance size to m5.large we could save half of this
cost.