readme: note about uploading secrets
mention that if the pipeline fail on initial run to upload the various
secrets.

you can't upload secrets until you have run the pipeline a bit (to
create the bucket?) and the pipeline can't run cleanly until you have
uploaded secrets, the reality is that you will run the pipeline, see it
fail, then upload secrets and re-run.