Update VagrantFile to work with modern vagrant
We had let this bitrot and it required using an old version of Vagrant.
Unfortunatly there is the need to put in a hack due to this issue on the
vagrant-aws provider
https://github.com/mitchellh/vagrant-aws/issues/566. I suspect that this
provider is now abandoned as there is even a PR open to fix it
https://github.com/mitchellh/vagrant-aws/pull/575. I propose that we
look to remove vagrant in the not too distant future.