Fix build CI
We've made some work to incorporate persistent dev environments into our
lives. This brought some difficulties with the secrets and has been
addressed in e14faf5678abda425968de208afd6c81984ddce1.

This however, breaks the weird separation in our Build CI system, where
the DEPLOY_ENV is suppose to be `build` but it should live in the `ci`
namespace... Story for another time...

This change, is suppose to fix this confusion and always set the correct
concourse URL for the build CI despite what environment variable is
responsible for this...