concourse: Bump number of web instances
Raising the number of workers will do us good for executing different set of tasks. But we also need more capacity in our brain to manage and control all of the available resources.

It would be OK to set the number at two, but in the same time, I don't want to revisit this for foreseeable future, and I want to make sure we have enough free play, when managing the instances.