Fix warning for undeclared db maintenance windows
During the `cyber-terraform` job these warning occurred.

```
Warning: Value for undeclared variable

The root module does not declare a variable named
"concourse_db_maintenance_window" but a value was found in file
"paas-bootstrap/terraform/dev.tfvars". To use this value, add a "variable"
block to the configuration.

Using a variables file to set an undeclared variable is deprecated and will
become an error in a future release. If you wish to provide certain "global"
settings to all configurations in your organization, use TF_VAR_...
environment variables to set these instead.

Warning: Value for undeclared variable

The root module does not declare a variable named "bosh_db_maintenance_window"
but a value was found in file "paas-bootstrap/terraform/dev.tfvars". To use
this value, add a "variable" block to the configuration.

Using a variables file to set an undeclared variable is deprecated and will
become an error in a future release. If you wish to provide certain "global"
settings to all configurations in your organization, use TF_VAR_...
environment variables to set these instead.
```

Proactively fix the warnings by adding them to globals.tf so it is not an issue in the future