Ensure correct Github Oauth credentials are picked for shared dev envs
There are special conditions surrounding the setting of secrets for dev envs:
1. If the MAKEFILE_ENV_TARGET environment variable is 'dev', the path in the
   password store will be 'github.com/concourse/${DEPLOY_ENV}/...'

2. If the MAKEFILE_ENV_TARGET environment variable is 'dev' AND the
   GITHUB_PASSWORD_STORE_DIR environment variable is NOT '${HOME}/.paas-pass',
   then the path in the secret store will be
   'github.com/concourse{$MAKEFILE_ENV_TARGET}/...'

This is done to account for the fact we have two types of development
environments:
* Persistent, shared environments like 'dev01'
* Short-lived, personal environments, such as those used for penetration
  testing.

These two types of dev environments store their Github OAuth credentials in
different places.  The former keeps them in the shared 'paas-credentials'
repository, and the latter are kept in an engineers personal store for the
lifetime of the environment.