Use 'create-swap-delete' strategy for Concourse VMs
When rolling Concourse VMs, under the default strategy, the
Concourse worker node responsible for running the job which tells
Bosh to roll the VMs is destroyed mid-job.

This is OK because the job of rolling the VMs is being performed by
the Bosh director, and the worker was only tailing the logs.

However, it does cause Concourse to think the job errored because it
can no longer contact the worker.

In a previous commit we told Concourse to attempt the job twice, but
we found that on the second attempt it was unable to start the job
because there was no worker available.

Changing the VM update strategy to 'creat-swap-delete' will ensure a
new worker is created and registered with Concourse before the old
one is thrown away. Therefore the second attempt should find a
functioning worker.