Run 'deploy-concourse' step on web co-located worker
The 'deploy-concourse' job can cause the worker on which it's running to go
away, resulting in the job erroring because Concourse can't contact it. To
counteract this, we run it on a worker which lives on the same machine as the
web node.

Therefore if the job gets interrupted, it's because the web node is going away.
This combined with the 2 attempts configuration from a previous commit should
mean that the job is completed on the second attempt, which will be kicked off
when the web node returns.