Update the ssh command for 'ssh_bosh'
Before, we were ssh'ing directly to the bosh ip, which is difficult to
set up ssh_config for.

This way, we ssh to `paas_bosh_ssh', but set the hostname to the IP we
are ssh'ing to.

~/.ssh/config can then be set up for this fake hostname.
```
Host paas_bosh_ssh
	User xxxx
	IdentityFile %d/.ssh/xyz.pub
```

Specifying the user for the SSH to `${USER}` is unnecessary, as this is the
default behaviour of SSH, and makes it impossible to override with ssh_config.
Also, as `$USER` is not immutable, it's not a security improvement.

This is an improvement which allows for cases where there are a large number
of ssh keys in the agent ie. if using [1Password's agent](https://developer.1password.com/docs/ssh/).
In this case, if the needed key is too far down the list, the attempts with
invalid keys leads to an authfail before the correct one is tried.