Scale up the BOSH director database to db.m5.large
We recently encountered lock-up of the BOSH director with the immediate
symptom being loss of metrics. In addition to the loss of metrics we see
max load on the BOSH director and high load, swap and max connections on
the BOSH RDS database.
We hypothesize that the lock-up is caused by database latency in the
first place.
Abstracting bosh_db_instance_class with a default of db.t3.small. The
default is in effect for all non production environments.