drop-bosh-credentials-for-rotation: preserve credhub_encryption_password
if we rotate this, credhub is unable to access its database and
the next pipeline run will fail. if you really *do* want to
rotate the credhub_encryption_password you're either going to
lose the contents of the credhub database, in which case you may
as well just delete the whole vars file and start again, or
you'll have to perform some deeper magic than this job is capable
of to re-encrypt them with a new key.