Change CONCOURSE_INSTANCE_TYPE to c5a.xlarge
We appear to be CPU bound on the concourse instances in `dev` instances,
which is causing acceptance tests to run really slowly, and fail due to
timeouts (#183781224).

These issues are not present in `staging` and `prod` environments, and
my working theory is that this is because these environments are running
`m5.xlarge` concourse instances, while `dev` is running `m5.large`.

Increasing the number of vCPUs on the concourse instances in `dev` in
line with those in `prod` should reduce failures and therefore reduce
dev time taken 'babysitting' pipeline runs when combined with some
optimisations in paas-cf (alphagov/paas-cf#3047).

We have two options to choose from to increase the vCPU count:
horizontal or vertical scaling. I have chosen vertical scaling
because a large amount of time is `wasted` by concourse shipping assets
and working directories between workers. Larger individual nodes means
that this is done less often.

EC2 Model details as-per 22/12/2022 in eu-west-1:

| Model          | vCPU (#) | RAM (GiB) | Cost per hour (USD) |
|----------------|----------|-----------|---------------------|
| m5.large       | 2        | 8         | 0.107               |
| **c5a.xlarge** | 4        | 8         | 0.172               |
| c5.xlarge      | 4        | 8         | 0.192               |
| m5.xlarge      | 4        | 16        | 0.214               |
| c5a.2xlarge    | 8        | 16        | 0.344               |
| c5.2xlarge     | 8        | 16        | 0.384               |

As we are CPU bound, it makes sense to keep the available RAM to a
minimum to keep costs down, meaning we should look at the `c5` range.
Currently, we have no reserved instances available in the `dev` account,
so we have some more freedom to make our choice of models.

As far as I can tell, we are unlikely to observe any real difference
between `c5` (3.6GHz Intel Xeon Platinum 8000) and `c5a` (3.3GHz AMD
EPYC 7002) instances for our workloads, so I have chosen `c5a.xlarge`
as the new `CONCOURSE_INSTANCE_TYPE`. If we need to bump these again,
`c5a.2xlarge` should probably be considered.