Enable DNS hostnames on VPC
Enabling DNS hostnames in the VPC will cause Amazon to assign a DNS name to
each EC2 instance with a public IP address [1]. We have historically not had
this enabled because it defaults to being off.

However, we will need it enabled to use certain types of AWS VPC endpoints.
Enabling it should have no negative affects on the rest of the VPC.

[1] https://docs.aws.amazon.com/vpc/latest/userguide/vpc-dns.html#vpc-dns-hostnames