bosh db: upgrade postgresql 11 -> 13
this requires some trickery and will require some cleanup that can't
happen while we still have envs that haven't done the upgrade.

it's also possible that the new engine_version will be rejected in
different environments which happen to be on different minor versions
of postgres 11, where 13.10 isn't a valid upgrade target. 13.10 is a
valid upgrade target for 11.19.